feat: Implement link shortener module
This commit adds a new module called `shorten.rs` in the `generator` directory. The `shorten.rs` module contains the implementation of a link shortener. It includes a trait `Shortener` and a struct `LinkShortener` that implements the trait. The `LinkShortener` struct uses a `NamesRepository` and a `NameGeneratorTrait` to generate shortened names for input URLs. It also provides input validation on the links.

The purpose of this commit is to provide a link shortening functionality that can be used to generate unique and shortened names for URLs.